Item DB Routine checks, Integrity improvements and Item Creation order (#261)

+++ inventory/items/models/Order/OrderModels.py
@@ -1,5 +1,6 @@
 from django.contrib.auth.models import User
 from django.db import models
+from django.db import transaction
 
 from inventory.items.globals import action_choices, action_reasons
 
@@ -31,13 +32,28 @@ class Order(models.Model):
     action = models.CharField(max_length=50, choices=action_choices)
     reason = models.CharField(max_length=50, choices=action_reasons)
     date = models.DateTimeField(auto_now_add=True)
-    user = models.ForeignKey(User, on_delete=models.CASCADE, related_name="orders")
+    user = models.ForeignKey(User, on_delete=models.RESTRICT, related_name="orders")
     other_info = models.CharField(max_length=100, null=True, blank=True)
+    is_reverted = models.BooleanField(default=False)
 
     def revert_order(self):
-        for item in self.order_items.all():
-            item.revert_order_item()
-        self.delete()
+        with transaction.atomic():
+            for item in self.order_items.all():
+                item.revert_order_item()
+            self.is_reverted = True
+            self.save()
+
+            # delete order if not linked to any order items
+            if self.order_items.count() == 0:
+                self.delete()
+
+    def check_valid_state(self):
+        """
+        Checks the following:
+        1. Order reason is Others, then other_info should not be None
+        """
+        if self.reason == "Others" and self.other_info is None:
+            raise Exception("Order reason is Others but other_info is None")
 
     def __str__(self) -> str:
         return f"{self.action}, {self.reason}, {self.date}, {self.user}"
